requirements.txt
flask
flask-cors
flasgger
python-dotenv
requests
beautifulsoup4
pymongo
redis
